# go/go/src/ipsap/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "IPSAP Service APIs",
        "contact": {},
        "license": {},
        "version": "1.0"
    },
    "basePath": "/api/v1.0",
    "paths": {
...
        "/platform/admin-user": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "관리자 정보\n우선 관리자 연락처 관리 정보에서 서비스 관리자로만 호출하면됨!\nfilter.user_auth = 9  플랫폼 관리자\nfilter.user_auth = 10 서비스 관리자\nfilter.user_auth = 11 시스템 관리자",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Platform"
                ],
                "summary": "관리자 정보",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "1",
                        "name": "filter.user_auth",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {}
                }
            },
            "patch": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "관리자 정보 수정\nuser_seq 포함해서 데이터 만들어야됨! (연락처, 이메일은 필수 값이 아님!)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Platform"
                ],
                "summary": "관리자 정보 수정",
                "parameters": [
                    {
                        "description": "param",
                        "name": "param",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/model.platformAdminUserPatchModel"
                        }
                    }
                ],
                "responses": {
                    "200": {}
                }
            }
        },
        "/platform/institution": {
            "patch": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Platform관리자가 기관정보 수정\nuser_seq institution_seq를 추가로 param을 만들어 줘야함!\nuserArr에 행정간사데이터를 담아서 보내준다!",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "multipart/form-data"
                ],
                "tags": [
                    "Platform"
                ],
                "summary": "Platform관리자가 기관정보 수정",
                "parameters": [
                    {
                        "type": "file",
                        "description": "기관 로고 파일",
                        "name": "logo_file",
                        "in": "formData"
                    },
                    {
                        "type": "file",
                        "description": "사업자 등록증 파일",
                        "name": "business_file",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "기관정보 수정(Json String 형태)",
                        "name": "param",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {}
                }
            }
        },
        "/platform/other-institution/{user_seq}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "타 기관 유저 리스트",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Platform"
                ],
                "summary": "타 기관 유저 리스트",
                "parameters": [
                    {
                        "type": "string",
                        "description": "user_seq",
                        "name": "user_seq",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {}
                }
            }
        },
        "/platform/payment/cancel/institution": {
            "patch": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "관리자가 결제 취소 후 기관 서비스 이용 상태 변경",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Platform"
                ],
                "summary": "관리자가 결제 취소 후 기관 서비스 이용 상태 변경",
                "parameters": [
                    {
                        "description": "param",
                        "name": "param",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/model.InstitutionServiceChangeModel"
                        }
                    }
                ],
                "responses": {
                    "200": {}
                }
            }
        },
        "/platform/user": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "등록 회원 관리 리스트",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Platform"
                ],
                "summary": "등록 회원 관리 리스트",
                "responses": {
                    "200": {}
                }
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "회원 등록",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Platform"
                ],
                "summary": "회원 등록",
                "parameters": [
                    {
                        "description": "param",
                        "name": "param",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/model.platformUserRegisterModel"
                        }
                    }
                ],
                "responses": {
                    "200": {}
                }
            },
            "patch": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Platform관리자가 사용자 정보 수정",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Platform"
                ],
                "summary": "Platform관리자가 사용자 정보 수정",
                "parameters": [
                    {
                        "description": "param",
                        "name": "param",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/model.platformPatchUserInfoModel"
                        }
                    }
                ],
                "responses": {
                    "200": {}
                }
            }
        },
        "/platform/user-batch": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "회원 일괄 등록",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Platform"
                ],
                "summary": "회원 일괄 등록",
                "parameters": [
                    {
                        "description": "param",
                        "name": "param",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/model.platformUserRegisterBatchModel"
                        }
                    }
                ],
                "responses": {
                    "200": {}
                }
            }
        },
        "/platform/user/{user_seq}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "등록 회원 관리 정보",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Platform"
                ],
                "summary": "등록 회원 관리 정보",
                "parameters": [
                    {
                        "type": "string",
                        "description": "user_seq",
                        "name": "user_seq",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {}
                }
            },
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "등록 대기중인 사용자 삭제",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Platform"
                ],
                "summary": "등록 대기중인 사용자 삭제",
                "parameters": [
                    {
                        "type": "string",
                        "description": "user_seq",
                        "name": "user_seq",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {}
                }
            }
        },
        "/platform/user/{user_seq}/resend-msg": {
            "patch": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "등록 대기 상태인 유저 가입 메세지 재발송",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Platform"
                ],
                "summary": "등록 대기 상태인 유저 가입 메세지 재발송",
                "parameters": [
                    {
                        "type": "string",
                        "description": "user_seq",
                        "name": "user_seq",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {}
                }
            }
        },
        "/platform/user/{user_seq}/reset-password": {
            "patch": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "등록회원 관리 비밀번호 초기화",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Platform"
                ],
                "summary": "등록회원 관리 비밀번호 초기화",
                "parameters": [
                    {
                        "type": "string",
                        "description": "user_seq",
                        "name": "user_seq",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {}
                }
            }
        },
        "/platform/user/{user_seq}/withdraw": {
            "patch": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "회원 강제 탈퇴(실제 삭제 X 상태만 변경)\n전체 기관에서 탈퇴함\n각 기관의 행정간사에게 메일을 보내야함",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Platform"
                ],
                "summary": "회원 강제 탈퇴",
                "parameters": [
                    {
                        "type": "string",
                        "description": "user_seq",
                        "name": "user_seq",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {}
                }
            }
        },
...
    },
    "definitions": {
...
        "model.InstitutionServiceChangeModel": {
            "type": "object",
            "properties": {
                "institution_seq": {
                    "type": "integer"
                },
                "service_status": {
                    "type": "integer",
                    "example": 1
                }
            }
        },
...
        "model.platformAdminUserPatchModel": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string",
                    "example": "test@test.com"
                },
                "name": {
                    "type": "string",
                    "example": "홍길동"
                },
                "telno": {
                    "type": "string",
                    "example": "0212341234"
                },
                "user_seq": {
                    "type": "integer",
                    "example": 1
                }
            }
        },
        "model.platformPatchUserInfoModel": {
            "type": "object",
            "properties": {
                "agree_email": {
                    "type": "integer",
                    "example": 1
                },
                "agree_pri_open": {
                    "type": "integer",
                    "example": 1
                },
                "agree_sms": {
                    "type": "integer",
                    "example": 1
                },
                "dept": {
                    "type": "string"
                },
                "edu_course_num": {
                    "type": "string",
                    "example": "0"
                },
                "edu_date": {
                    "type": "string",
                    "example": "20210101"
                },
                "edu_institution": {
                    "type": "string",
                    "example": "건국대학교"
                },
                "institution_seq": {
                    "type": "integer"
                },
                "major_field": {
                    "type": "string",
                    "example": "수의학"
                },
                "phoneno": {
                    "type": "string",
                    "example": "021231234"
                },
                "position": {
                    "type": "string",
                    "example": "직급"
                },
                "user_seq": {
                    "type": "integer"
                },
                "user_status": {
                    "type": "integer"
                },
                "user_type": {
                    "type": "string"
                }
            }
        },
        "model.platformUserRegisterBatchModel": {
            "type": "object",
            "properties": {
                "userArr": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.platformUserRegisterModel"
                    }
                }
            }
        },
        "model.platformUserRegisterModel": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string",
                    "example": "test@test.com"
                },
                "institution_seq": {
                    "type": "integer",
                    "example": 1
                },
                "phoneno": {
                    "type": "string",
                    "example": "01012341234"
                },
                "user_type": {
                    "type": "string",
                    "example": "1,2"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "token",
            "in": "header"
        }
    }
}